Add a link to production version of VSCode plugin
It's hard to make your way from this repo to the plugin. We should make it easier to navigate by adding a link.

I originally was going to put the link in https://github.com/sveltejs/language-tools/blob/master/packages/svelte-vscode/README.md in https://github.com/sveltejs/language-tools/pull/76 but it appears that README is used directly in https://marketplace.visualstudio.com/items?itemName=svelte.svelte-vscode so then I decided that wasn't the best place to put it

I also am clarifying in this PR that this repo is for a VSCode language server and that language servers aren't generic across IDEs. This was confusing to me and others. E.g. see https://github.com/sveltejs/language-tools/issues/74#issuecomment-625913226