(feat) support default language of svelte-preprocess (#291)
* (feat) support default language of svelte-preprocess

`svelte-preprocess` has a feature for setting default languages so you don't have to write `lang="x"` each time. This adds support for this in style/script tags.

To accomplish this, parsing of the `svelte.config.js` was moved up to `Document`. Inside it, the `lang` attribute is set to the default language if there exists no `lang`/`type` attribute.