(fix) mark props as optional in ts when no typings given (#370)
Before this, `export let a = ''` would not be marked as optional because it had no type.
#347

Also fixes a bug where file type was incorrectly inferred as jsx which occurs when user uses the default languages feature of svelte-preprocess. Instead of computing the file type inside svelte2tsx, it is now passed as an option.