(fix) awaited type of promise or plain value (#439)
When awaiting the value of a type like Promise<string> | null, the type
wouldn't match PromiseLike. Hence the awaited value would still be
Promise<string> | null and not string | null. Reducing it to one type
declaration and using a conditional type to infer the awaited type
should do the trick.