(fix) only provide extension version of prettier plugin if there is no workspace version
Only provide our version of the svelte plugin if the user doesn't have one in the workspace already. If we did it, Prettier would - for some reason - use the workspace version for parsing and the extension version for printing, which could crash if the contract of the parser output changed.
(which it did from 1.1.0 to 1.1.1)