(feat) brand new tmLanguage grammar for Svelte
Replaces the current gramamr with a new one that is significantly more feature-rich. Despite this, it's only about a 100 lines longer than the original (in JSON form).
It is compiled from a `yaml` file. When I was making it, I just used `js-yaml` to compile it. An example of how to do this is found on the official VSCode syntax highlighting guide.