Transform if-blocks to ternaries
#619
This way the TypeScript control flow will be broken in less instances and nested if-blocks like
```
{#if a}
  {#if a.a}
     ..
  {/if}
{/if}
```
will not throw "a could be undefined" in the inner if anymore in strict mode.
This will not shut down all instances of control flow breakage, because for example slots/each/await are still transformed to lambda-functions.