(fix) fallback param for built-in transition/animate
Up to Svelte version 3.32.0, the built-in animate/transition functions have optional parameters, but according to its typings they were mandatory. To not show unnecessary type errors to those users, `{}` should be added as a fallback parameter if the user did not provide one.
It may be the case that someone has a custom animation with the same name that expects different parameters, but that possibility is far less likely.
#785