(feat) more robust parsing for mustache tags (#786)
In svelte2tsx, run a simple scan before sending the code to the svelte compiler. Right now the compiler errors on mustache tags that are incomplete and are for example ended by an operand like ".". This change allows the errored syntax to be emitted untouched, allowing better DX.
The behavior is toggleable and off by default so that `svelte-check` will still emit these errors.

#538