(perf) improve svelte-check performance (#872)
Because of the setTimeout in the debounceSameArg. The document update in LSAndTSDocResolver is actually done after all the diagnostic is done. Because of that, the typescript language service has to rebuild its type checker multiple times due to a new document being loaded in. Fix that by not debouncing the call on a new document. This makes svelte-check run a lot faster.