(fix) accessor creation without type or with rename (#895)
Prior to this, props would only get into setters if they had a type. This changes it so that all props are picked up and renamed exports are now also supported. This is achieved by reusing the exportedNames-map which removes the need for a dedicated setters prop on the render function return object.
#888