(fix) add esm flag to support node 12 < version < 12.17
Node versions prior to v12.17 need the `--experimental-modules` flag or the dynamic import won't work. This node version is used in VS Code versions prior to 1.55 and could be used by users setting the node executable to a node version below 12.17.

Remove this in mid 2022 and bump vs code minimum required version to 1.55