(fix) update document logic
Due to the refactoring in #917, documents were reread from disk every time when getting a snapshot. This resulted in worse performance and introduced inconsistencies because "openDocument" of the DocumentManager was called over and over again with potentially outdated content, resulting in buggy, seemingly random behavior.
This fix reverts that part of #917 which was accidentally introduced and only loads documents from disk if a snapshot doesn't already exist.