(perf) also cache unresolved modules (#1015)
This change also makes resolveModuleNames return early if the module to be resolved was already tried to be resolved but without success. Now, `undefined` is returned immediately in this case without trying to resolve again. Such unresolved resolutions are deleted whenever a new file is created. This increases performance especially on projects with many unresolved modules.

During this change it was made sure that SnapshotManager is now only called through the service. This also fixes a small bug where a TS/JS file was not removed from the module resolution after it was deleted.