(breaking) be more strict with required props (#1030)
This change now always marks properties as required if they have no initializer. Previously this was only the case when using TypeScript with strict mode. For this to work for both TS and JS files, the transformation in svelte2tsx was adjusted. 

#319
#847 

### This is a breaking change
The fix is to provide a default value for properties that are optional. That's either the specific value kind or `undefined`.
Example TS:
`export let optional: string;`
Becomes
`export let optional: string = '';` (or just `export let optional = ''`)
Example JS:
`export let optional;`
Becomes
`export let optional = undefined;`