(breaking) rename ambient __sveltets_-declarations
Rename all __sveltets_* ambient declarations to __sveltets_1_* . This way, we avoid the situation where multiple ambient versions of svelte2tsx are loaded and their declarations conflict each other. #1059 for an example bug that stems from it.
The other breaking change is that svelte2tsx no longer provides a default export, svelte2tsx now is a named export. This opens up the possibility to export other utility functions alongside it.