(breaking) adjust svelte2tsx dts generation
Didn't work for JS files with js doc as it produces invalid JS files (type definitions were present). This splits the generation up into a special JS-compliant format. For that, the ambient type definitions are now expected to be provided by the user. The type definitions were adjusted so that no type-artifacts remain in the output by forcing flat type results through more usage of the Expand type.