(perf) typescript language service size limit (#1194)
#965
Add a size limit for typescript language service similar to the one in the tsserver. When the size limit exceeds, ts language service goes into a reduced mode. Project-wide IntelliSense is disabled. It would only use the files opened in the client. Kind of like when you don't have a tsconfig.json file. When this happens a warning message will show up.

Note if you have a really big project and you have excluded everything needed but it still exceeds this limit, you can disable it with the disableSizeLimit compiler options in the tsconfig/jsconfig.