(fix) add open tag checks to html scanner
the html parser does treat a "<" tag inside the body of another tag as the start of another tag. This is a common situation inside #if control flow tags, so these were blanked out inside parseTag previously. But that blanking logic did not take place when parsed html nodes were passed in. Therefore moved the logic of stripping #if tags inside the preprocess scanner