(fix) ignore refs in generated tsx / non user code (#1276)
A lot of variables references currently end up in generated tsx but are not surrounded by "generated code" guards. As a result, the language servers currently emits references for locations (and refs / code) that don't exist in user's source code. It turns out those ghost references are petty easy to filter out because, as they don't match anything in user's code, they are source mapped back to 0 length ranges.