(fix) Silence diagnostics for no side effects variables in reactive blocks (#1286)
#1277

For diagnostics "Left side of comma operator is unused and has no side effects.", this code analyzes the AST to determine whether the errors point to a let variable in a reactive statement. In this case, we assume this is purposeful reactive block hinting and suppress the error. New diagnostics are emitted for each non let identifiers that exists in the span of the original diagnostic.

To fix the main case, an AST based isReactiveStatement util function is introduced. The "Unused label." warnings for $: of reactive statements used a weak heuristic, so the warning suppression strategy was switched to using the AST based function too.