(fix) store rework (#1531)
Reworks $store handling in the code base. Previously, svelte2tsx would transform $store access to something like (__get(store), $store). This is now no longer the case. This means we need to add some extra code to some of the intellisense features:

- if someone renames a $store, store also needs to be renamed and vice versa
- if someone wants to find references of a $store, we also need to find store references and vice versa
- if someone wants to go to the definition of a $store, we need to reroute that to store
- if someone wants to go to the implementation of a $store, we need to reroute that to store
diagnostics of wrong $store usages need to be handled

This means some more code overall, but I feel it's the better choice since we had so many edge cases with our transformation other time, and #1475 being an upstream bug we cannot solve ourselves but need to solve for the new transformation.