(feat) use moustache for bind, on:handler completion (#1676)
#1465
Use moustache for bind, on:handler completion instead of the current double quote. And if svelteStrictMode is set to true in the prettier config, use "{}".

So If you want to use "{}" instead of {}. Set svelteStrictMode to true in your prettier config file. If you don't have a prettier config file. You can also set it with the svelte.plugin.svelte.format.config.svelteStrictMode config.