(feat) fix-all quick fix (#1939)
#1885

There are two parts to this PR. First is supporting the combined quick for TypeScript. And Second part is to add our "import component" and "import store" fix on top of it. Handler is skipped since TypeScript 5.0 should have full support, and our custom code would only be backup for non-VSCode users still using < 5.0.

This is implemented by creating a copy of the file with the declared name of the store and component variables at the end of the script tags. This way, we don't have to deal with the "which is first" and import organization problems. Performance-wise, It's sometimes slower than the custom solution because TypeScript will rebuild and invalidate the cache. But it's more consistent. The original method might be very slow if there are a lot of missing imports.