(fix) highlight destructuring in const tag (#2015)
#2013.

Using the same approach for each tag destructuring. Because the } in the restructuring will be part of the source.ts matching. It won't count as the end of the @const tag. Different from each tag restructuring, this uses source.ts#object-binding-pattern-const and source.ts#array-binding-pattern-const to highlight the variable as variable.other.constant.ts.