fix: prevent crashes when ts-plugin config updates (#2091)
#2086

The crash happens because we clear the semantic cache when the config updates. And now TypeScript actively Debug.assert if the program object is the same as the last time they update it. But in most cases, the config update is because of the API we use to update the config after the plugin is started. The semantic cache clearing seems unnecessary, at least in most cases, so it should be fine if we remove it and should stop the crash.