chore: try fix flaky test (#2203)
There are two files that don't exist that were referenced in the test. We must rely on openClientDocument to open them. But It seems like there is a race condition somewhere that would close it after it's opened. I'm not really sure if this is the case since I can't reproduce it locally even if I run it at least 100 times.

If the file does exist, it'll be loaded as a project file the first time the ts service is created. And the openClientDocument will only be used as a failed safe. Hopefully, it will fix the problem. I also added a new test to test the language feature of the client file that doesn't exist in the file system. It's tested in a virtual system, so it should be more isolated and less likely to fail randomly.