chore: consolidate template walking logic
right now some of the logic of walking the template is within the svelte2tsx part, some in the htmlxtojsx part. This makes it hard to decide where to put new stuff and represents an arbitrary barrier for features that may need to work across both.

This PR therefore moves all the logic from the svelte2tsx part into the htmlxtojsx part.