feat: make use of loose Svelte parser and provide better intellisense
This adds support for the Svelte parser with its new loose mode and adjusts code paths to make use of it properly. As a result, intellisense should be a lot more useful in situations where code is in the middle of being typed and the Svelte file is in a broken state.