fix: invalidate project file cache and handle watcher race condition (#2779)
#2773

This delays the file reading in svelte-check --watch until we're sure the new file is a project file. I still added a file existence check before reading, but it might be more of a fail-safe because the chance of it being deleted at this point is much smaller.

While checking this, I also found that the new file watch is broken because it now caches the tsconfig parsing result. So I changed the "read the config again" logic to the one language server used to invalidate the project files cache.