feat: workspace symbols support (#2769)
#2375.
Adding support in the Svelte language server so that you don't need to open ts/js files to use this feature. You can trigger this feature with Ctrl + T and search for symbol names. Some of the handling is also synced to the typescript-plugin so that it's more consistent, which also means vscode can better deduplicate the results from both tsserver and Svelte language server.